Change default recursive flag to false when purging files.
I had a chat with someone in Stackpath and it seems the recursive=true is causing issues with the processing speed. It also will result in slower processing of the purges requested by the purge_files function. So the recursive parameter should be used only when it is really needed.